ml-service/requirements.txt
fastapi
uvicorn
pydantic
scikit-learn
pandas
numpy
openai
python-dotenv
